Fix wrap and props:
