+++ app/components/base/property-list/index.ts
@@ -1,5 +1,5 @@
 export * from "./property-list.component";
 export * from "./property-list.module";
-export * from "./text-property.component";
+export * from "./text-property";
 export * from "./table-property";
 export * from "./property-group.component";
