+++ app/components/base/property-list/text-property/index.ts
@@ -1,0 +1,1 @@
+export * from "./text-property.component";
