+++ app/components/base/property-list/text-property/text-property.html
@@ -1,0 +1,9 @@
+<bl-property-field (mouseenter)="showClipboard = true" (mouseleave)="showClipboard = false">
+    <div propertyLabel>{{label}}</div>
+    <bl-property-content class="property-content-box" [wrap]="wrap" [title]="value">
+        {{value}}
+    </bl-property-content>
+    <div propertyClipboard [hidden]="!showClipboard" class="clipboard" *ngIf="copyable">
+        <i class="fa fa-clipboard" (click)="copyToClipBoard()"></i>
+    </div>
+</bl-property-field>
